chore: add run instructions to KStreams scheduling operations tutorial

+++ kafka-streams-schedule-operations/kstreams/src/main/java/io/confluent/developer/LoginTime.java
@@ -1,4 +1,4 @@
 package io.confluent.developer;
 
-public record LoginTime(long logInTime, String userId, String appId) {
+public record LoginTime(int logInTime, String userId, String appId) {
 }
